requirements.txt
youtube_dl
imgurpython
moviepy
praw
datetime
PyNaCl
opuslib
psutil

aiohttp>=3.6.0
websockets>=6.0,<7.0
discord.py